+++ snakeway-core/src/conf/types/specification/bind/rate_limiter_filter.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Default, Serialize, Clone)]
 pub struct RateLimiterFilterSpec {
-    pub max_connections_per_second: f64,
-    pub reaction_interval_in_seconds: u64,
+    pub max_connections_per_second: u16,
+    pub reaction_interval_in_seconds: u16,
 }
